chore: fix failing tests
