+++ src/lib.rs
@@ -2,6 +2,7 @@ mod compaction;
 mod entry;
 mod error;
 mod indexer;
+mod iter;
 pub mod log;
 mod manifest;
 mod meta;
